Define a call to a weird function

weirds function are wrapper around other packages anomalies detection routines. weird function is intended to be used when invoking stranger so that several anomalies detection routines are used at one time.

weird(method, ...)

Arguments

method

- weird method to be used, check list of available methods by using weirds_list

...

- additional parameters to be passed to weird method

Value

list with class weirdSpecs to be used as value when using the parameter tuneList of stranger function.

For every weird function, possible parameters in ... should be checked by looking to the help of the underlying function. Some special common parameters are info to be set to TRUE to have information about the method, inclusing the name and package of the main underlying function. Another common parameter is colname that allow to overide default prefix name used for the outcome of the computation.

Methods specific parameters, additional parameters may be used (for instance simplify for _knn_ method or type for kmeans. See the vignette describing the list of available _weirds_ methods in base package.