weirdness wrappers for available anomalies detection methods

Those wrapper are mainly used internally by other functions and the recommended way is to start your detection process by using one of: strange, stranger or lucky_odds functions.

weird_knn(data = NULL, info = FALSE, colname = NULL, simplify = "mean",
  ...)

weird_lof(data = NULL, info = FALSE, colname = NULL, ...)

weird_autoencode(data = NULL, info = FALSE, colname = NULL, ...)

weird_abod(data = NULL, info = FALSE, colname = NULL, ...)

weird_pcout(data = NULL, info = FALSE, colname = NULL, ...)

weird_isofor(data = NULL, info = FALSE, colname = NULL, ...)

weird_kmeans(data = NULL, info = FALSE, colname = NULL, type = "means",
  ...)

weird_mahalanobis(data = NULL, info = FALSE, colname = NULL, ...)

weird_randomforest(data = NULL, info = FALSE, colname = NULL, ...)

Arguments

data
data pre-processed with crazyfy function

info
logical: either run the function (default) or simply return some information about it

colname
character: optional prefix to be used to prepare the naming of the generated column. All weird methods do have a default prefix (applied when colname is NULL). Note that generated output column name will still have some information added to reflect values used for some key parameters.

simplify
character of function - used by knn to aggregates distances to nearest k points.

…
additional parameters to be passed to weird method, for instance k for knn

Details

For each weird wrapper, please refer to the documentation of the underlying function in its package.